+++ src/main/kotlin/app/revanced/patcher/PackageMetadata.kt
@@ -1,15 +1,14 @@
-package app.revanced.patcher.data
+package app.revanced.patcher
 
 import brut.androlib.apk.ApkInfo
 
 /**
  * Metadata about a package.
  */
-class PackageMetadata {
+class PackageMetadata internal constructor(internal val apkInfo: ApkInfo) {
     lateinit var packageName: String
         internal set
+
     lateinit var packageVersion: String
         internal set
-
-    internal lateinit var apkInfo: ApkInfo
 }