+++ src/test/kotlin/app/revanced/patcher/usage/resource/patch/ExampleResourcePatch.kt
@@ -2,10 +2,7 @@ package app.revanced.patcher.usage.resource.patch
 
 import app.revanced.patcher.annotation.Description
 import app.revanced.patcher.annotation.Name
-import app.revanced.patcher.annotation.Version
 import app.revanced.patcher.data.ResourceContext
-import app.revanced.patcher.patch.PatchResult
-import app.revanced.patcher.patch.PatchResultSuccess
 import app.revanced.patcher.patch.ResourcePatch
 import app.revanced.patcher.patch.annotations.Patch
 import app.revanced.patcher.usage.resource.annotation.ExampleResourceCompatibility
@@ -15,9 +12,8 @@ import org.w3c.dom.Element
 @Name("example-resource-patch")
 @Description("Example demonstration of a resource patch.")
 @ExampleResourceCompatibility
-@Version("0.0.1")
 class ExampleResourcePatch : ResourcePatch {
-    override fun execute(context: ResourceContext): PatchResult {
+    override fun execute(context: ResourceContext) {
         context.xmlEditor["AndroidManifest.xml"].use { editor ->
             val element = editor // regular DomFileEditor
                 .file
@@ -29,7 +25,5 @@ class ExampleResourcePatch : ResourcePatch {
                     "exampleValue"
                 )
         }
-
-        return PatchResultSuccess()
     }
 }